refactor: extractDataAsAccount가 모든 UserProfile 타입을 받도록 변경
